+++ src/main/resources/changelog.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+
+## [1.3.0]
+
+### Added
+- hungry zombie
+
+### Changed
+- soggy/fungal zombies will now leave a trail of slime
+
 ## [1.2.0]
 
 ### Changed
@@ -37,7 +47,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - rich zombie
 - fungal zombie
 
-[1.2.0]: https://github.com/BarchamMal/The-Zombie-Mod/commit/
+[1.3.0]: https://github.com/BarchamMal/The-Zombie-Mod/commit/
+[1.2.0]: https://github.com/BarchamMal/The-Zombie-Mod/commit/be61318820ef02f40c38157e6baf2fc546411b7d
 [1.1.0]: https://github.com/BarchamMal/The-Zombie-Mod/commit/bf10833ff8a99d90331f2067d0a70447b7d27a84
 [1.0.1]: https://github.com/BarchamMal/The-Zombie-Mod/commit/d34c0d784b00c65a85fb91dfeeb9f1c0da54d515
 [1.0.0]: https://github.com/BarchamMal/The-Zombie-Mod/commit/9ff8d79c982e1c56747bdd1802c4ffc14515dda2